Add tests / Handle encoding and decoding
