+++ OmiseGOTests/CodingTests/DecodeTests.swift
@@ -25,6 +25,33 @@ class DecodeTests: XCTestCase {
         return try Data(contentsOf: fixtureFileURL)
     }
 
+    func testSuccessfullyDecodesANumberWith38Digits() {
+        do {
+            let jsonData = try self.jsonData(withFileName: "bigint_valid")
+            let decodedData =  try self.jsonDecoder.decode(BigIntDummy.self, from: jsonData)
+            XCTAssertEqual(decodedData.value.description, "99999999999999999999999999999999999999")
+        } catch let thrownError {
+            XCTFail(thrownError.localizedDescription)
+        }
+    }
+
+    func testFailsToDecodeANumberWith39Digits() {
+        do {
+            let jsonData = try self.jsonData(withFileName: "bigint_invalid")
+            let decodedData =  try self.jsonDecoder.decode(BigIntDummy.self, from: jsonData)
+            XCTAssertEqual(decodedData.value.description, "999999999999999999999999999999999999991")
+        } catch let error as DecodingError {
+            switch error {
+            case .dataCorrupted(let context):
+                XCTAssertEqual(context.debugDescription, "Invalid number")
+            default:
+                XCTFail("Should raise a data corrupted error")
+            }
+        } catch _ {
+            XCTFail("Should raise a decoding error")
+        }
+    }
+
     func testCustomDateDecodingStrategy() {
         let jsonDecoder = JSONDecoder()
         jsonDecoder.dateDecodingStrategy = .custom({return try dateDecodingStrategy(decoder: $0)})
