+++ OmiseGOTests/FixtureTests/Fixtures/objects/bigint_invalid.json
@@ -1,0 +1,3 @@
+{
+    "value": 999999999999999999999999999999999999991
+}
