+++ OmiseGOTests/FixtureTests/Fixtures/objects/bigint_valid.json
@@ -1,0 +1,3 @@
+{
+    "value": 99999999999999999999999999999999999999
+}
